efw-protocols: add implementation for protocol about robot guitar
This commit adds implementation for protocol about robot guitar. The batch
of code is moved from internal transaction module to new module.
